# openapi.json
{
  "components": {
    "schemas": {
...
      "EventExpenseCreate": {
        "additionalProperties": false,
        "properties": {
          "amount": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Amount"
          },
          "category_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "individual_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Individual Id"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "type": {
            "const": "expense",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "occurred_at",
          "type",
          "amount"
        ],
        "title": "EventExpenseCreate",
        "type": "object"
      },
      "EventIncomeCreate": {
        "additionalProperties": false,
        "properties": {
          "amount": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Amount"
          },
          "category_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "individual_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Individual Id"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "type": {
            "const": "income",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "occurred_at",
          "type",
          "amount"
        ],
        "title": "EventIncomeCreate",
        "type": "object"
      },
      "EventInventoryCreate": {
        "additionalProperties": false,
        "properties": {
          "adjustment": {
            "$ref": "#/components/schemas/InventoryAdjustment"
          },
          "category_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "individual_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Individual Id"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "quantity": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Quantity"
          },
          "type": {
            "const": "inventory",
            "title": "Type",
            "type": "string"
          },
          "unit": {
            "$ref": "#/components/schemas/Unit"
          }
        },
        "required": [
          "occurred_at",
          "type",
          "adjustment",
          "quantity",
          "unit"
        ],
        "title": "EventInventoryCreate",
        "type": "object"
      },
      "EventObservationCreate": {
        "additionalProperties": false,
        "properties": {
          "category_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "individual_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Individual Id"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "type": {
            "const": "observation",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "occurred_at",
          "type"
        ],
        "title": "EventObservationCreate",
        "type": "object"
      },
      "EventProductionCreate": {
        "additionalProperties": false,
        "properties": {
          "category_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "individual_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Individual Id"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "quantity": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Quantity"
          },
          "type": {
            "const": "production",
            "title": "Type",
            "type": "string"
          },
          "unit": {
            "$ref": "#/components/schemas/Unit"
          }
        },
        "required": [
          "occurred_at",
          "type",
          "quantity",
          "unit"
        ],
        "title": "EventProductionCreate",
        "type": "object"
      },
...
      "EventReproductiveCreate": {
        "additionalProperties": false,
        "properties": {
          "category_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "individual_id": {
            "title": "Individual Id",
            "type": "integer"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "type": {
            "const": "reproductive",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "occurred_at",
          "individual_id",
          "type"
        ],
        "title": "EventReproductiveCreate",
        "type": "object"
      },
      "EventType": {
        "enum": [
          "production",
          "expense",
          "income",
          "observation",
          "reproductive",
          "acquisition",
          "mortality",
          "inventory"
        ],
        "title": "EventType",
        "type": "string"
      },
      "EventUpdate": {
        "additionalProperties": false,
        "properties": {
          "adjustment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InventoryAdjustment"
              },
              {
                "type": "null"
              }
            ]
          },
          "amount": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "category_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "individual_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Individual Id"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "occurred_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurred At"
          },
          "payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload"
          },
          "quantity": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "unit": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Unit"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "EventUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InventoryAdjustment": {
        "enum": [
          "increment",
          "decrement",
          "reset"
        ],
        "title": "InventoryAdjustment",
        "type": "string"
      },
      "InventoryBalance": {
        "properties": {
          "asset_id": {
            "title": "Asset Id",
            "type": "integer"
          },
          "balances": {
            "items": {
              "$ref": "#/components/schemas/InventoryBalanceRow"
            },
            "title": "Balances",
            "type": "array"
          }
        },
        "required": [
          "asset_id",
          "balances"
        ],
        "title": "InventoryBalance",
        "type": "object"
      },
...
      "Page_EventRead_": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/EventRead"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "Page[EventRead]",
        "type": "object"
      },
...
      "Unit": {
        "enum": [
          "g",
          "kg",
          "lb",
          "t",
          "ml",
          "l",
          "gal",
          "unit",
          "dozen",
          "head"
        ],
        "title": "Unit",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "Ovejitas API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/farms/{farm_id}/assets/{asset_id}/events": {
      "get": {
        "operationId": "events_list_events",
        "parameters": [
          {
            "in": "path",
            "name": "farm_id",
            "required": true,
            "schema": {
              "title": "Farm Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "integer"
            }
          },
          {
            "description": "Search notes",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search notes",
              "title": "Q"
            }
          },
          {
            "description": "e.g. -occurred_at",
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "e.g. -occurred_at",
              "title": "Sort"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/EventType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "in": "query",
            "name": "category_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category Id"
            }
          },
          {
            "in": "query",
            "name": "individual_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Individual Id"
            }
          },
          {
            "in": "query",
            "name": "adjustment",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/InventoryAdjustment"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Adjustment"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_EventRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List events for an asset",
        "tags": [
          "events"
        ]
      },
      "post": {
        "description": "Body is a discriminated union on `type`. Reproductive events require `kind=animal` + `individual_id`. Individuals require `mode=individual` and must belong to this asset. Categories must match the event type.",
        "operationId": "events_create_event",
        "parameters": [
          {
            "in": "path",
            "name": "farm_id",
            "required": true,
            "schema": {
              "title": "Farm Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "discriminator": {
                  "mapping": {
                    "expense": "#/components/schemas/EventExpenseCreate",
                    "income": "#/components/schemas/EventIncomeCreate",
                    "inventory": "#/components/schemas/EventInventoryCreate",
                    "observation": "#/components/schemas/EventObservationCreate",
                    "production": "#/components/schemas/EventProductionCreate",
                    "reproductive": "#/components/schemas/EventReproductiveCreate"
                  },
                  "propertyName": "type"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/EventProductionCreate"
                  },
                  {
                    "$ref": "#/components/schemas/EventExpenseCreate"
                  },
                  {
                    "$ref": "#/components/schemas/EventIncomeCreate"
                  },
                  {
                    "$ref": "#/components/schemas/EventObservationCreate"
                  },
                  {
                    "$ref": "#/components/schemas/EventReproductiveCreate"
                  },
                  {
                    "$ref": "#/components/schemas/EventInventoryCreate"
                  }
                ],
                "title": "Data"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create an event",
        "tags": [
          "events"
        ]
      }
    },
    "/api/v1/farms/{farm_id}/assets/{asset_id}/events/balance": {
      "get": {
        "description": "Returns the derived on-hand balance per (asset, unit), computed from INVENTORY events: sum of increments minus decrements since the most recent reset. Only meaningful for assets with kind=material.",
        "operationId": "events_asset_inventory_balance",
        "parameters": [
          {
            "in": "path",
            "name": "farm_id",
            "required": true,
            "schema": {
              "title": "Farm Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryBalance"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Current on-hand inventory for a material asset",
        "tags": [
          "events"
        ]
      }
    },
    "/api/v1/farms/{farm_id}/assets/{asset_id}/events/{event_id}": {
      "delete": {
        "description": "Events emitted by an action (those carrying a `payload.source`) cannot be deleted here. Deleting an inventory event is rejected if it would drive the stock balance negative.",
        "operationId": "events_delete_event",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "farm_id",
            "required": true,
            "schema": {
              "title": "Farm Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete an event",
        "tags": [
          "events"
        ]
      },
      "get": {
        "operationId": "events_get_event",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "farm_id",
            "required": true,
            "schema": {
              "title": "Farm Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get one event",
        "tags": [
          "events"
        ]
      },
      "patch": {
        "description": "`type` is immutable. Events emitted by an action (those carrying a `payload.source`) cannot be edited here \u2014 edit them via their action. Fields must match the event's type, and inventory edits are re-checked against the stock balance.",
        "operationId": "events_update_event",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "farm_id",
            "required": true,
            "schema": {
              "title": "Farm Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update an event",
        "tags": [
          "events"
        ]
      }
    },
...
  }
}